+++ resources/views/admin/school/show.blade.php
@@ -12,18 +12,20 @@
 
                         <a href="{{ url('/admin/school') }}" title="Back"><button class="btn btn-default btn-md"><i class="fa fa-arrow-left" aria-hidden="true"></i> Back</button></a>
                         <a href="{{ url('/admin/school/' . $school->id . '/edit') }}" title="Edit School"><button class="btn btn-primary btn-md"><i class="fa fa-pencil-square-o" aria-hidden="true"></i> Edit</button></a>
-                        {!! Form::open([
-                            'method'=>'DELETE',
-                            'url' => ['admin/school', $school->id],
-                            'style' => 'display:inline'
-                        ]) !!}
-                            {!! Form::button('<i class="fa fa-trash-o" aria-hidden="true"></i> Delete', array(
-                                    'type' => 'submit',
-                                    'class' => 'btn btn-danger btn-md',
-                                    'title' => 'Delete School',
-                                    'onclick'=>'return confirm("Confirm delete?")'
-                            ))!!}
-                        {!! Form::close() !!}
+                        @can('delete school')
+                            {!! Form::open([
+                                'method'=>'DELETE',
+                                'url' => ['admin/school', $school->id],
+                                'style' => 'display:inline'
+                            ]) !!}
+                                {!! Form::button('<i class="fa fa-trash-o" aria-hidden="true"></i> Delete', array(
+                                        'type' => 'submit',
+                                        'class' => 'btn btn-danger btn-md',
+                                        'title' => 'Delete School',
+                                        'onclick'=>'return confirm("Confirm delete?")'
+                                ))!!}
+                            {!! Form::close() !!}
+                        @endcan
                         <br/>
                         <br/>
